backend/requirements.txt
fastapi[all]
python-socketio[client]
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
